+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This changelog consists of the bug & security fixes and new features being inclu
 
 * Optimized RMA-related queries and introduced a return period column in the order items table.
 
+* Fixed issues with language switching in the installation wizard and corrected PHP configuration texts.
+
 * #11100 [fixed] - Fixed an issue where updating the return window rule affected previously placed orders.
 
 ## **v2.4.0-beta1 (9th of February 2026)** - **Release**